+++ frontend/src/components/Home/Home_main.tsx
@@ -6,13 +6,11 @@ import "src/styles/style.css";
 import "src/styles/home.css";
 
 function Home() {
-
     return (
         <div className="mainContainerRow">
             <div>
                 <div className="h2">Home</div>
                 <Login />
-               
             </div>
         </div>
     );
