+++ frontend/src/components/LeaderBoard/Leaderboard_main.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import UserTable from "../shared/UserTable";
 import RankNumberColumn from "./RankNumberColumn";
-import { fetchGetSet } from "src/functions/fetchers";
+import { fetchGetSet } from "src/functions/utils";
 
 // DTO
 import { UserProfileDTO } from "user-dto";
